OWN 트레이너 식단 운동 간단 팁(데이터기반으로 간단로직)구현

+++ src/main/java/com/ownlife/service/MealServiceImpl.java
@@ -126,7 +126,7 @@ public class MealServiceImpl implements MealService {
 
             for (Map<String, Object> foodData : foods) {
                 Long foodId = Long.valueOf(String.valueOf(foodData.get("foodId")));
-                double count = Integer.parseInt(String.valueOf(foodData.get("count")));
+                double count = Double.parseDouble(String.valueOf(foodData.get("count")));
 
                 addMeal(memberId, date, mealType, foodId, count);
             }
